shuffle vec/array/tuple futures/streams for benchmark (with a fixed seed)
